% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countHeatmap.R
\name{countHeatmap}
\alias{countHeatmap}
\title{Produce a HiC Count Heatmap}
\usage{
countHeatmap(data)
}
\arguments{
\item{data}{A TopDomData object.}
}
\value{
A \link[ggplot2:ggplot]{ggplot2::ggplot} object.
}
\description{
Produce a HiC Count Heatmap
}
\seealso{
See \link{TopDom} for an example.
}
