% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggCountHeatmap.R
\name{ggCountHeatmap}
\alias{ggCountHeatmap}
\title{Produce a Count Heatmap}
\usage{
ggCountHeatmap(data, transform, colors, ...)
}
\arguments{
\item{data}{A TopDomData object.}

\item{transform}{A function applied to the counts prior to generating
heatmap colors.}

\item{colors}{A named list to control to color scale.}

\item{\ldots}{Not used.}
}
\value{
A \link[ggplot2:ggplot]{ggplot2::ggplot} object.
}
\description{
Produce a Count Heatmap
}
\seealso{
See \link{TopDom} for an example.
}
\author{
Henrik Bengtsson.
}
