% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggCountHeatmap.R
\name{ggDomain}
\alias{ggDomain}
\title{Add a Topological Domain to a Count Heatmap}
\usage{
ggDomain(td, delta = 0.04, size = 2, vline = 0, color = "#666666")
}
\arguments{
\item{td}{A single-row TopDomData object.}

\item{delta}{Relative distance to heatmap.}

\item{size, color}{The thickness and color of the domain line.}

\item{vline}{Additional absolute distance to heatmap.}
}
\value{
A \link[ggplot2:geom_segment]{ggplot2::geom_segment} object to be added to the count heatmap.
}
\description{
Add a Topological Domain to a Count Heatmap
}
