% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legacy.R
\name{legacy}
\alias{legacy}
\title{Easy Access to the Original TopDom 0.0.1 and 0.0.2 Implementations}
\usage{
legacy(version = c("0.0.1", "0.0.2"))
}
\arguments{
\item{version}{A version string.}
}
\value{
An environment containing the legacy TopDom API.
}
\description{
Easy Access to the Original TopDom 0.0.1 and 0.0.2 Implementations
}
\examples{
TopDom::legacy("0.0.2")$TopDom
TopDom::legacy("0.0.1")$Detect.Local.Extreme

}
