% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{stop_if_not}
\alias{stop_if_not}
\title{Asserts the Truth of R Expressions}
\usage{
stop_if_not(...)
}
\arguments{
\item{\dots}{Zero or more \R expressions to be asserted to be TRUE.}
}
\value{
Nothing.
}
\description{
Asserts the Truth of R Expressions
}
\details{
A bare bone, faster version of \link[base:stopifnot]{base::stopifnot}.
}
\keyword{internal}
