% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsetByRegion.R
\name{subsetByRegion}
\alias{subsetByRegion}
\title{Subset a TopDomData Object by Region}
\usage{
subsetByRegion(data, region, margin = 1/2)
}
\arguments{
\item{data}{A TopDomData object.}

\item{region}{A TopDom domain (a data.frame).}

\item{margin}{An non-negative numeric specifying the additional margin
extracted around the domain.
If \code{margin < 1}, then the size of the margin is relative
to the size of the domain.}
}
\value{
A TopDomData object.
}
\description{
Subset a TopDomData Object by Region
}
\author{
Henrik Bengtsson.
}
