% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TopDom.R
\name{TopDom}
\alias{TopDom}
\title{Identify topological domains for given Hi-C contact matrix}
\usage{
TopDom(data, window.size, outFile = NULL, statFilter = TRUE, ...)
}
\arguments{
\item{data}{A TopDomData object or the pathname to a normalized
Hi-C contact matrix file as read by \link{readHiC}.}

\item{window.size}{The number of bins to extend (as a non-negative integer).
Recommended range is in {5, ..., 20}.}

\item{outFile}{(optional) The filename without extension of the two result
files optionally produced.}

\item{statFilter}{logical, ...}

\item{...}{Additional arguments passed to \link{readHiC}.}
}
\value{
A named list with data.frame elements \code{binSignal}, \code{domain},
and \code{bed}.
The \code{bed} data frame has columns \code{chrom}, \code{chromStart}, \code{chromEnd},
and \code{name}.
If argument \code{outFile} is non-\code{NULL}, then the three elements returned
are also written to tab-delimited files with file names
\file{<outFile>.binSignal}, \file{<outFile>.domain}, and
\file{<outFile>.bed}, respectively.  None of the files have row names,
and all but the BED file have column names.
}
\description{
Identify topological domains for given Hi-C contact matrix
}
\author{
Hanjun Shin, Harris Lazaris, and Gangqing Hu.
\R package, help, and code refactorization by Henrik Bengtsson.
}
