% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countsPerRegion.R
\name{countsPerRegion}
\alias{countsPerRegion}
\title{Calculates Counts per Region in a TopDomData Object}
\usage{
countsPerRegion(data, regions)
}
\arguments{
\item{data}{A TopDomData object.}

\item{regions}{TopDom regions (a data.frame), e.g. domains.}
}
\value{
A numeric vector of length \code{nrow(regions)}.
}
\description{
Calculates Counts per Region in a TopDomData Object
}
\author{
Henrik Bengtsson.
}
