% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggCountHeatmap.R
\name{ggDomain}
\alias{ggDomain}
\title{Add a Topological Domain to a Count Heatmap}
\usage{
ggDomain(td, dx = NULL, delta = 0.04, vline = 0, size = 2, color = "#666666")
}
\arguments{
\item{td}{A single-row data.frame.}

\item{dx, delta, vline}{Absolute distance to heatmap.
If \code{dx = NULL} (default), then \code{dx = delta * w + vline} where \code{w} is
the width of the domain.}

\item{size, color}{The thickness and color of the domain line.}
}
\value{
A \link[ggplot2:geom_segment]{ggplot2::geom_segment} object to be added to the count heatmap.
}
\description{
Add a Topological Domain to a Count Heatmap
}
