% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlapScores.R
\name{overlapScores}
\alias{overlapScores}
\title{Calculates Overlap Scores between Two Sets of Topological Domains}
\usage{
overlapScores(a, b, ...)
}
\arguments{
\item{a, b}{Topological domain sets 'A' and 'B', where set 'A' contains
D_A domains and set 'B' D_B domains.}

\item{\ldots}{...}
}
\description{
Calculates Overlap Scores between Two Sets of Topological Domains
}
\seealso{
\link{TopDom}.
}
