% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlapScores.R
\name{overlapScores}
\alias{overlapScores}
\title{Calculates Overlap Scores between Two Sets of Topological Domains}
\usage{
overlapScores(a, b, ..., debug = getOption("TopDom.debug", FALSE))
}
\arguments{
\item{a, b}{Topological domain (TD) sets \eqn{A} and \eqn{B}
as returned by \code{\link[=TopDom]{TopDom()}}.}

\item{debug}{If \code{TRUE}, debug output is produced.}

\item{\ldots}{...}
}
\value{
Returns a named list of class \code{TopDomOverlapScores}, where the names
correspond to the chromosomes in domain set \eqn{A}.
Each of these chromosome elements contains a named list of elements
\code{best_score} (\eqn{D_{A,c}} numerics in \eqn{[0,1]}) and
`best_sets' (\eqn{D_{A,c}} list of index vectors), where
\eqn{D_{A,c}} is the number of TDs in chromosome \eqn{c} in set \eqn{A}.
}
\description{
Calculates Overlap Scores between Two Sets of Topological Domains
}
\details{
The \emph{overlap score}, \eqn{overlap(a_i, B')}, represents how well topological
domain (TD) \eqn{a_i} in set \eqn{A} overlap with a \emph{consecutive} subset
\eqn{B'} of TDs in \eqn{B}.
For each TD \eqn{a_i}, the \emph{best match} \eqn{B'_max} is identified, that
is, the \eqn{B'} subset that maximize \eqn{overlap(a_i, B')}.
For exact definitions, see Page 8 in Shin et al. (2016).

Note that the overlap score is an asymmetric score, which means that
\code{overlapScores(a, b) != overlapScores(b, a)}.
}
\examples{
if (require("TopDomData")) {
  path <- system.file("exdata", package = "TopDomData", mustWork = TRUE)

  ## Original count data
  pathname <- file.path(path, "nij.chr10.gz")
  data <- readHiC(pathname, chr = "chr10", binSize = 40e3)
  print(data)

  ## Find topological domains using TopDom method for two window sizes
  tds_5 <- TopDom(data, window.size = 5L)
  tds_6 <- TopDom(data, window.size = 6L)

  ## Overlap scores (in both directions)
  overlap_56 <- overlapScores(tds_5, tds_6)
  print(overlap_56)
  
  overlap_65 <- overlapScores(tds_6, tds_5)
  print(overlap_65)
}
}
\references{
\itemize{
\item Shin et al.,
TopDom: an efficient and deterministic method for identifying
topological domains in genomes,
\emph{Nucleic Acids Research}, 44(7): e70, April 2016.
doi: 10.1093/nar/gkv1505,
PMCID: \href{https://www.ncbi.nlm.nih.gov/pmc/articles/PMC4838359/}{PMC4838359},
PMID: \href{https://www.ncbi.nlm.nih.gov/pubmed/26704975}{26704975}.
}
}
\seealso{
\link{TopDom}.
}
\author{
Henrik Bengtsson - based on the description in Shin et al. (2016).
}
