% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TopDom.R
\name{TopDom}
\alias{TopDom}
\title{Identify topological domains for given Hi-C contact matrix}
\usage{
TopDom(matrix.file, chr = NULL, binSize = NULL, window.size,
  outFile = NULL, statFilter = TRUE)
}
\arguments{
\item{matrix.file}{The pathname of a normalize Hi-C contact matrix file
stored as a whitespace-delimited file.  See below for details.
Also a gzip-compressed file can be used.}

\item{window.size}{The number of bins to extend (as a non-negative integer).
Recommended range is in {5, ..., 20}.}

\item{outFile}{(optional) The filename without extension of the two result
files optionally produced.}

\item{statFilter}{logical, ...}
}
\value{
A named list with data.frame elements \code{binSignal}, \code{domain},
and \code{bed}.
The \code{bed} data frame has columns \code{chrom}, \code{chromStart}, \code{chromEnd},
and \code{name}.
If argument \code{outFile} is non-\code{NULL}, then the three elements returned
are also written to tab-delimited files with file names
\file{<outFile>.binSignal}, \file{<outFile>.domain}, and
\file{<outFile>.bed}, respectively.  None of the files have row names,
and all but the BED file have column names.
}
\description{
Identify topological domains for given Hi-C contact matrix
}
\author{
Hanjun Shin, Harris Lazaris, and Gangqing Hu.
\R package, help, and code refactorization by Henrik Bengtsson.
}
