% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TopDom.R
\name{TopDom}
\alias{TopDom}
\title{Identify topological domains for given Hi-C contact matrix}
\usage{
TopDom(matrix.file, window.size, outFile = NULL, statFilter = TRUE)
}
\arguments{
\item{matrix.file}{The pathname of a normalize Hi-C contact matrix file
stored as a whitespace-delimited file.  See below for details.}

\item{window.size}{The number of bins to extend (as a non-negative integer).
Recommended range is in {5, ..., 20}.}

\item{outFile}{(optional) The filename without extension of the two result
files optionally produced.}

\item{statFilter}{logical, ...}
}
\value{
A named list with data.frame elements \code{binSignal}, \code{domain},
and \code{bed}.
The \code{bed} data frame has columns \code{chrom}, \code{chromStart}, \code{chromEnd},
and \code{name}.
If argument \code{outFile} is non-\code{NULL}, then the three elements returned
are also written to tab-delimited files with file names
\file{<outFile>.binSignal}, \file{<outFile>.domain}, and
\file{<outFile>.bed}, respectively.  None of the files have row names,
and all but the BED file have column names.
}
\description{
Identify topological domains for given Hi-C contact matrix
}
\section{Format of HiC contact-matrix file}{

The contact-matrix file should be a whitespace-delimited text file with
neither row names nor column names.  The content should be a N-by-(3+N)
table where the first three columns correspond to \code{chr} (string),
\code{from.coord} (integer position), and \code{to.coord} (integer position).
These column defines the genomic location of the N Hi-C bins (in order).
The last N columns should contain normalized contact counts (float) such
that element (r, 3+c) in this table corresponds to count (r,c) in the
normalized contact matrix.

If an N-by-(4+N) table, then the first column is assumed to contain an
\code{id} (integer), and everything else as above.

Example:
\preformatted{
  chr10       0   40000  0 0 0 0 ...
  chr10   40000   80000  0 0 0 0 ...
  chr10   80000  120000  0 0 0 0 ...
  chr10  120000  160000  0 0 0 0 ...
  ...
}
}

\author{
Hanjun Shin, Harris Lazaris, and Gangqing Hu.
\R package, help, and code refactorization by Henrik Bengtsson.
}
