% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TopDom.R
\name{TopDom}
\alias{TopDom}
\title{Identify topological domains for given Hi-C contact matrix}
\usage{
TopDom(matrix.file, window.size, outFile = NULL, statFilter = TRUE)
}
\arguments{
\item{window.size}{integer, number of bins to extend.}

\item{matrixFile}{string, matrixFile Address,
\itemize{
\item Format = {chromosome, bin start, bin end, N numbers normalized value }
\item N * (N + 3), where N is the number of bins
}}

\item{out_binSignal}{string, binSignal file address to write}

\item{out_ext}{string, ext file address to write}
}
\description{
Identify topological domains for given Hi-C contact matrix
}
\author{
Hanjun Shin, Harris Lazaris, and Gangqing Hu.
R package, help and code refactoring by Henrik Bengtsson.
}
