% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TopDom.R
\name{TopDom}
\alias{TopDom}
\title{Identify topological domains for given Hi-C contact matrix}
\usage{
TopDom(data, window.size, outFile = NULL, statFilter = TRUE, ...)
}
\arguments{
\item{data}{A TopDomData object or the pathname to a normalized
Hi-C contact matrix file as read by \link{readHiC}.}

\item{window.size}{The number of bins to extend (as a non-negative integer).
Recommended range is in {5, ..., 20}.}

\item{outFile}{(optional) The filename without extension of the two result
files optionally produced.}

\item{statFilter}{logical, ...}

\item{...}{Additional arguments passed to \link{readHiC}.}
}
\value{
A named list with data.frame elements \code{binSignal}, \code{domain},
and \code{bed}.
The \code{bed} data frame has columns \code{chrom}, \code{chromStart}, \code{chromEnd},
and \code{name}.
If argument \code{outFile} is non-\code{NULL}, then the three elements returned
are also written to tab-delimited files with file names
\file{<outFile>.binSignal}, \file{<outFile>.domain}, and
\file{<outFile>.bed}, respectively.  None of the files have row names,
and all but the BED file have column names.
}
\description{
Identify topological domains for given Hi-C contact matrix
}
\examples{
if (require("TopDomData")) {
  path <- system.file("exdata", package = "TopDomData", mustWork = TRUE)

  ## Original count data
  pathname <- file.path(path, "nij.chr10.gz")
  data <- readHiC(pathname, chr = "chr10", binSize = 40e3)
  print(data)

  ## Find topological domains using TopDom method
  fit <- TopDom(data, window.size = 5L)
  str(fit$domain)
}
}
\references{
\itemize{
\item Shin et al.,
TopDom: an efficient and deterministic method for identifying
topological domains in genomes,
\emph{Nucleic Acids Res.} 2016 Apr 20; 44(7): e70., 2015.
doi: 10.1093/nar/gkv1505,
PMCID: \href{https://www.ncbi.nlm.nih.gov/pmc/articles/PMC4838359/}{PMC4838359},
PMID: 26704975.
\item Shin et al., \R script \file{TopDom_v0.0.2.R}, 2017.
\url{http://zhoulab.usc.edu/TopDom/}.
\item Shin et al., TopDom Manual, 2016-07-08.
\url{http://zhoulab.usc.edu/TopDom/TopDom\%20Manual_v0.0.2.pdf}.
}
}
\author{
Hanjun Shin, Harris Lazaris, and Gangqing Hu.
\R package, help, and code refactorization by Henrik Bengtsson.
}
